rename hooke sphere operators
