+++ src/sphere/hooke_sphere.cpp
@@ -1,13 +1,13 @@
 /*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
+   Licensed to the Apache Software Foundation (ASF) under one
+   or more contributor license agreements.  See the NOTICE file
+   distributed with this work for additional information
+   regarding copyright ownership.  The ASF licenses this file
+   to you under the Apache License, Version 2.0 (the
+   "License"); you may not use this file except in compliance
+   with the License.  You may obtain a copy of the License at
 
-  http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing,
 software distributed under the License is distributed on an
@@ -15,7 +15,9 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
-*/
+ */
+//#pragma xstamp_cuda_enable //! DO NOT REMOVE THIS LINE
+
 #include <exanb/core/operator.h>
 #include <exanb/core/operator_slot.h>
 #include <exanb/core/operator_factory.h>
@@ -29,13 +31,13 @@ under the License.
 #include <exaDEM/compute_hooke_force.h>
 #include <exaDEM/interaction/interaction.hpp>
 #include <exaDEM/interaction/grid_cell_interaction.hpp>
+#include <exaDEM/interaction/classifier.hpp>
+#include <exaDEM/interaction/classifier_for_all.hpp>
 #include <exaDEM/shape/shapes.hpp>
 #include <exaDEM/shape/shape_detection.hpp>
 #include <exaDEM/shape/shape_detection_driver.hpp>
 #include <exaDEM/mutexes.h>
 #include <exaDEM/drivers.h>
-#include <exaDEM/shape/shapes.hpp>
-#include <exaDEM/shape/shape_detection.hpp>
 #include <exaDEM/hooke_sphere.h>
 
 namespace exaDEM
@@ -44,155 +46,81 @@ namespace exaDEM
   using namespace sphere;
 
   template<typename GridT , class = AssertGridHasFields< GridT, field::_radius >>
-    class ComputeHookeInteractionSphere : public OperatorNode
+    class ComputeHookeClassifierSphereGPU : public OperatorNode
   {
     // attributes processed during computation
     using ComputeFields = FieldSet< field::_vrot, field::_arot >;
     static constexpr ComputeFields compute_field_set {};
 
-    ADD_SLOT( GridT                       , grid          , INPUT_OUTPUT , REQUIRED );
-    ADD_SLOT( GridCellParticleInteraction , ges           , INPUT_OUTPUT , DocString{"Interaction list"} );
-    ADD_SLOT( HookeParams                 , config        , INPUT        , REQUIRED ); // can be re-used for to dump contact network
-    ADD_SLOT( HookeParams                 , config_driver , INPUT        , OPTIONAL ); // can be re-used for to dump contact network
-    ADD_SLOT( mutexes                     , locks         , INPUT_OUTPUT );
-    ADD_SLOT( double                      , dt            , INPUT        , REQUIRED );
-    ADD_SLOT( bool                        , symetric      , INPUT        , REQUIRED , DocString{"Activate the use of symetric feature (contact law)"} );
-    ADD_SLOT( Drivers                     , drivers       , INPUT        , DocString{"List of Drivers"});
-    ADD_SLOT( double                      , rcut_max      , INPUT_OUTPUT , 0.0 );
-    ADD_SLOT( vector_t<size_t>            , idxs          , INPUT_OUTPUT , DocString{"List of non empty cells"});
+    ADD_SLOT( GridT       , grid              , INPUT_OUTPUT , REQUIRED );
+    ADD_SLOT( HookeParams , config            , INPUT , REQUIRED , DocString{"Hooke parameters for sphere interactions"}); // can be re-used for to dump contact network
+    ADD_SLOT( HookeParams , config_driver     , INPUT , OPTIONAL , DocString{"Hooke parameters for drivers, optional"}); // can be re-used for to dump contact network
+    ADD_SLOT( double      , dt                , INPUT , REQUIRED , DocString{"Time step value"});
+    ADD_SLOT( bool        , symetric          , INPUT , REQUIRED , DocString{"Activate the use of symetric feature (contact law)"});
+    ADD_SLOT( Drivers     , drivers           , INPUT , DocString{"List of Drivers {Cylinder, Surface, Ball, Mesh}"});
+    ADD_SLOT( Classifier  , ic                , INPUT_OUTPUT , DocString{"Interaction lists classified according to their types"} );
+
 
     public:
 
     inline std::string documentation() const override final
     {
       return R"EOF(
+
                 )EOF";
     }
 
     inline void execute () override final
     {
-      Drivers empty;
-      Drivers& drvs =  drivers.has_value() ? *drivers : empty;
-      const double rcut = config->rcut;
-      *rcut_max = std::max( *rcut_max , rcut );
-
       if( grid->number_of_cells() == 0 ) { return; }
 
-      const auto cells = grid->cells();
-      auto & cell_interactions = ges->m_data;
-      const HookeParams params = *config;
-      const double time = *dt;
-      mutexes& locker = *locks;
-      auto& indexes = *idxs;
-      HookeParams hkp_drvs;
+      Drivers* drvs =  drivers.get_pointer();
+      auto* cells = grid->cells();
+      const HookeParams hkp = *config;
+      HookeParams hkp_drvs{};
+
       if ( drivers->get_size() > 0 &&  config_driver.has_value() )
       {
         hkp_drvs = *config_driver;
-        *rcut_max = std::max( *rcut_max , hkp_drvs.rcut );
       }
 
-      // TODO later refactoring
+      const double time = *dt;
+      auto& classifier = *ic;
+
+      hooke_law_driver<Cylinder> cyli;
+      hooke_law_driver<Surface>  surf;
+      hooke_law_driver<Ball>     ball;
+      hooke_law_stl stlm = {};
+
+      cudaDeviceSynchronize();
+
       if(*symetric)
-      {
-        const hooke_law<true> sph;
-        const exaDEM::sphere::hooke_law_stl stl = {};
-        const exaDEM::sphere::hooke_law_driver<Cylinder> cyl;
-        const exaDEM::sphere::hooke_law_driver<Surface> surf;
-        const exaDEM::sphere::hooke_law_driver<Ball>    ball;
-        size_t idxs_size = onika::cuda::vector_size( indexes );
-        size_t *idxs_data = onika::cuda::vector_data( indexes ); 
-
-#pragma omp parallel for schedule(guided)
-        for( size_t ci = 0 ; ci < idxs_size ; ci ++ )
-        {
-          size_t current_cell = idxs_data[ci];  
-
-          auto& interactions = cell_interactions[current_cell];
-          const unsigned int data_size = onika::cuda::vector_size( interactions.m_data );
-          exaDEM::Interaction* const __restrict__ data_ptr = onika::cuda::vector_data( interactions.m_data ); 
-
-          for( size_t it = 0; it < data_size ; it++ )
-          {
-            Interaction& item = data_ptr[it];
-
-            if(item.type == 0) // sphere-sphere
-            {
-              sph(item, cells, params, time, locker);
-            }
-            else if(item.type == 4) // cylinder
-            {
-              cyl(item, cells, drvs, hkp_drvs, time, locker);
-            }
-            else if(item.type == 5) // surface
-            {
-              surf(item, cells, drvs, hkp_drvs, time, locker);
-            }
-            else if(item.type == 6) // ball
-            {
-              ball(item, cells, drvs, hkp_drvs, time, locker);
-            }
-            else if(item.type >= 7 && item.type <= 9) // stl
-            {
-              stl(item, cells, drvs, hkp_drvs, time, locker);
-            }
-          }
-        }
+			{
+        hooke_law<true> sph;
+        run_contact_law(parallel_execution_context(), 0, classifier, sph, cells, hkp, time);  
       }
       else
       {
-        const hooke_law<false> sph;
-        const exaDEM::sphere::hooke_law_stl stl = {};
-        const exaDEM::sphere::hooke_law_driver<Cylinder> cyl;
-        const exaDEM::sphere::hooke_law_driver<Surface> surf;
-        const exaDEM::sphere::hooke_law_driver<Ball>    ball;
-        size_t idxs_size = onika::cuda::vector_size( indexes );
-        size_t *idxs_data = onika::cuda::vector_data( indexes ); 
-
-#pragma omp parallel for schedule(guided)
-        for( size_t ci = 0 ; ci < idxs_size ; ci ++ )
-        {
-          size_t current_cell = idxs_data[ci];  
-
-          auto& interactions = cell_interactions[current_cell];
-          const unsigned int data_size = onika::cuda::vector_size( interactions.m_data );
-          exaDEM::Interaction* const __restrict__ data_ptr = onika::cuda::vector_data( interactions.m_data ); 
-
-          for( size_t it = 0; it < data_size ; it++ )
-          {
-            Interaction& item = data_ptr[it];
-
-            if(item.type == 0) // sphere-sphere
-            {
-              sph(item, cells, params, time, locker);
-            }
-            else if(item.type == 4) // cylinder
-            {
-              cyl(item, cells, drvs, hkp_drvs, time, locker);
-            }
-            else if(item.type == 5) // surface
-            {
-              surf(item, cells, drvs, hkp_drvs, time, locker);
-            }
-            else if(item.type == 6) // ball
-            {
-              ball(item, cells, drvs, hkp_drvs, time, locker);
-            }
-            else if(item.type >= 7 && item.type <= 9) // stl
-            {
-              stl(item, cells, drvs, hkp_drvs, time, locker);
-            }
-          }
-        }
+        hooke_law<false> sph;
+        run_contact_law(parallel_execution_context(), 0, classifier, sph, cells, hkp, time);  
       }
+      run_contact_law(parallel_execution_context(), 4, classifier, cyli, cells, drvs->ptr<Cylinder>(), hkp_drvs, time);  
+      run_contact_law(parallel_execution_context(), 5, classifier, surf, cells, drvs->ptr<Surface>(), hkp_drvs, time);  
+      run_contact_law(parallel_execution_context(), 6, classifier, ball, cells, drvs->ptr<Ball>(), hkp_drvs, time);  
+      for(int type = 7 ; type <= 9 ; type++)
+      {
+        run_contact_law(parallel_execution_context(), type, classifier, stlm, cells, drvs, hkp_drvs, time);  
+      }
+      cudaDeviceSynchronize();
     }
   };
 
-  template<class GridT> using ComputeHookeInteractionSphereTmpl = ComputeHookeInteractionSphere<GridT>;
+  template<class GridT> using ComputeHookeClassifierGPUTmpl = ComputeHookeClassifierSphereGPU<GridT>;
 
   // === register factories ===  
   CONSTRUCTOR_FUNCTION
   {
-    OperatorNodeFactory::instance()->register_factory( "hooke_sphere", make_grid_variant_operator< ComputeHookeInteractionSphereTmpl > );
+    OperatorNodeFactory::instance()->register_factory( "hooke_sphere", make_grid_variant_operator< ComputeHookeClassifierGPUTmpl > );
   }
 }
